Add rough design doc and TUF client for milestone 1
Added a design doc for "tuf-minimal" design, matching
https://github.com/jku/playground-tuf-minimal .
How design docs are stored should probably change (maybe a separate
directory for the repository designs matching the different milestones)
but currently everything is dumped in the same docs dir.

Added a client implementation into tuf/client/: it's 90% same as
baseline/client/. My current thinking is that we can leave baseline
as is and just update the tuf/ implementation as we progress through the
milestones.

Modified the baseline files to better match the current tuf design.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>