signer: Tweak uri management
The goal was to keep all unrecognized_fields handling inside
SignerRepository, but the issue with x-playground-online-uri
is that when we create keys in _get_online_input, we then
want to compare them to existing keys (to recognise when
changes are made from the case where no changes were necessary)...
so the created keys must have the unrecognised field in them already.

so remove the separate URIs from OnlineConfig, and embed them
in the key to begin with.