Use newer repository metadata accessors
* slightly fewer magic strings
* slightly simpler code
* much better annotation coverage

This still leaves a few cases in the repository side where
"self.open(rolename)" is used: when we're doing verification we need the
whole metadata, not just the signed part. The solution to this might be
to expose more verification api as part of the repo -- like
"self.verify(rolename)" which would figure out the delegator on its own.