Targets management version 4
Redesign targets management once more:
* Users still make target content changes (outside of the tool)
* Let repository do the actual metadata changes during signing-event
  (the method is called status() but signing_event_update() might be
  better)
* Repository also documents the target changes at high level in the
  signing event
* Users sign normally: "playground-sign <EVENT>"

This makes the UX much more neat and does not complicate the repository
too much.

The expected targets metadata change in e2e tests is because the
PlaygroundRepository does not figure out version numbers as smartly as
SignerRepository does, and ends up bumping versions more than needed: this
could be fixed but it's not critical targets metadata.

This change means "playground-status" now makes commits and pushes them
to signing event branch. This means the commands name is now not very
descriptive. Changing it should not be disruptive but I did not want to
grow this PR, so leaving that for later.