signer: Do not sign during close
Signing in close creates all sorts of edge cases we don't need to
handle: Just add modified roles to unsigned list, and expect the CLI
to sign explicictly. This also fixes the issue of having to sign
multiple times in one signing event.

Note that root is a special case since the signing keys are stored in
root itself.