status: Show known good root signer counts first
In a way the current valid signers are the more important ones when
listing root signer counts: the new proposed signers must reach
threshold too but that number should not make us trust this proposed
content.

The message should be made clearer as well but I could find a way to do
that without making it very long.