snapshot: Always publish a new repo version
Previously we only published a new repository artifact if the snapshot
process actually created new metadata. This left an edge case: if only
root was modified in a signing event, then snapshot process does not
create new metadata -- but we would still like to publish the repository
after that.

Always publish a repository artifact if snapshot process executes.
This likely leads to a few unnecessary repository uploads but it's
better than no uploads when needed.

Leave the "generated" variable in the action for future improvements,
just set it permanently to "true".

Fixes #101